+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.7.0] - 2026-07-03
+
 ### Fixed
 
 - **Command palette (`⌘⇧P`) felt unresponsive.** It didn't clear the previous